refactor: break assistant message into component and adjust alignment
